refactor(buildgen): split css.go into planning, scoping, and minify files
css.go had grown to 1,059 lines / 54 functions mixing three unrelated
responsibilities. Split along natural seams into focused files:

- css.go (636): CSS planning — discover/plan page/layout/component CSS,
  content hashing, output paths
- css_scope.go (335): component CSS scoping — keyframe/animation/selector
  rewriting
- css_minify.go (104): dependency-free CSS minifier

Both extracted blocks are pure string-to-string transforms with no
dependency on the planning types, so this is behavior-preserving. No
package boundary changes. Full buildgen suite passes; gofmt/vet clean.
